Renamed the NullModelsFoundException to EntityModelsNotFoundException for sustainability reasons Made the OEM Number nullable (things like accessories don't have it) Added PartSaveRequest to entity conversion in the part mapper Re-aligned the service to match the test expectations

+++ src/main/java/dev/ioannis/anemosparts/services/impl/PartServiceImpl.java
@@ -4,15 +4,15 @@ import dev.ioannis.anemosparts.domain.PartDto;
 import dev.ioannis.anemosparts.domain.requests.PartSaveRequest;
 import dev.ioannis.anemosparts.domain.responses.PartFindAllResponse;
 import dev.ioannis.anemosparts.entities.Model;
-import dev.ioannis.anemosparts.entities.OemNumber;
 import dev.ioannis.anemosparts.entities.Part;
-import dev.ioannis.anemosparts.exceptions.NullModelsFoundException;
+import dev.ioannis.anemosparts.exceptions.EntityModelsNotFoundException;
 import dev.ioannis.anemosparts.exceptions.OemNumberDoesNotExistException;
 import dev.ioannis.anemosparts.mappers.PartMapper;
 import dev.ioannis.anemosparts.repositories.ModelRepo;
 import dev.ioannis.anemosparts.repositories.OemRepo;
 import dev.ioannis.anemosparts.repositories.PartRepo;
 import dev.ioannis.anemosparts.services.PartService;
+import jakarta.persistence.EntityNotFoundException;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -36,49 +36,49 @@ public class PartServiceImpl implements PartService {
 
     @Override
     public PartDto save(PartSaveRequest request) {
-        var part = new Part();
-        part.setName(request.getName());
-        part.setPartNumber(request.getPartNumber());
-        part.setOemNumber(OemNumber.builder().number(request.getOemNumber()).build());
-        part.setPrice(request.getPrice());
-        part.setDescription(request.getDescription());
-        part.setModels(request.getModelIds().stream().map(modelId -> Model.builder().id(modelId).build()).toList());
+        if(request.getModelIds() == null || request.getModelIds().isEmpty()) {
+            throw new IllegalArgumentException("Save request is missing models");
+        }
+
+        var part = mapper.toEntity(request);
 
         return mapper.toDto(save(part));
     }
 
     @Override
     public PartDto update(Long id, PartSaveRequest request) {
-        var part = new Part();
+        if(request.getModelIds() == null || request.getModelIds().isEmpty()) {
+            throw new IllegalArgumentException("Save request is missing models");
+        }
+
+        var part = mapper.toEntity(request);
         part.setId(id);
-        part.setName(request.getName());
-        part.setPartNumber(request.getPartNumber());
-        part.setOemNumber(OemNumber.builder().number(request.getOemNumber()).build());
-        part.setPrice(request.getPrice());
-        part.setDescription(request.getDescription());
-        part.setModels(request.getModelIds().stream().map(modelId -> Model.builder().id(modelId).build()).toList());
 
         return mapper.toDto(save(part));
     }
 
     protected Part save(Part part) {
         if(modelRepo.findAllById(part.getModels().stream().map(Model::getId).toList()).size() != part.getModels().size()) {
-            throw new NullModelsFoundException("");
+            throw new EntityModelsNotFoundException("");
         }
 
         part.setModels(modelRepo.findAllById(part.getModels().stream().map(Model::getId).toList()));
 
-        if(oemRepo.existsByNumber(part.getOemNumber().getNumber())) {
-            throw new OemNumberDoesNotExistException("");
-        }
+        if(part.getOemNumber() != null) {
+            if(oemRepo.existsByNumber(part.getOemNumber().getNumber()))
+                part.setOemNumber(oemRepo.findByNumber(part.getOemNumber().getNumber()));
 
-        part.setOemNumber(oemRepo.findByNumber(part.getOemNumber().getNumber()));
+            else throw new OemNumberDoesNotExistException("");
+        }
 
         return partRepo.save(part);
     }
 
     @Override
     public void delete(long id) {
-        partRepo.deleteById(id);
+        if(partRepo.existsById(id)) {
+            partRepo.deleteById(id);
+        }
+        else throw new EntityNotFoundException("Could not find part with id: " + id + " to delete");
     }
 }
